Add transformation to concrete patterns; Refactoring
