+++ org.emoflon.ibex.common/META-INF/MANIFEST.MF
@@ -1,16 +1,18 @@
 Manifest-Version: 1.0
 Bundle-SymbolicName: org.emoflon.ibex.common;singleton:=true
-Bundle-Name: %pluginName
+Bundle-Name: eMoflon::IBeX Common
 Bundle-Version: 1.0.0.qualifier
 Bundle-ManifestVersion: 2
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: IBeXLanguage,
  IBeXLanguage.impl,
- IBeXLanguage.util
+ IBeXLanguage.util,
+ org.emoflon.ibex.common.utils
 Bundle-Vendor: %providerName
 Bundle-Localization: plugin
 Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.emf.ecore;visibility:=reexport,
+ org.eclipse.emf.ecore.xmi,
  org.junit
 Bundle-ClassPath: .
