+++ org.emoflon.ibex.common/build.properties
@@ -5,5 +5,6 @@ bin.includes = .,\
                META-INF/,\
                plugin.xml,\
                plugin.properties
-source.. = gen/
+source.. = gen/,\
+           src/
 output.. = bin/
